fix(cache): improve the caching methods
- Improve terminology
- Add logging
- Prevent calling Reminders.deleteMany when there are none to delete
- Add notion of "Offline cache" (doesn't use Discord API, called during
the Client initialization), and "Online Cache" (uses Discord API, called
when the bot is connected to the Gateway (= is ready)).
- Add a lock to ensure Online Cache is always called after Offline Cache

+++ src/tasks/reminders.ts
@@ -15,7 +15,8 @@ export class ReminderTask extends Task {
       date: { $lte: Date.now() - 7 * 24 * 60 * 60 * 1000 },
     });
 
-    await Reminder.deleteMany({ _id: { $in: remindersToRemove.map(r => r._id) } });
+    if (remindersToRemove.length > 0)
+      await Reminder.deleteMany({ _id: { $in: remindersToRemove.map(r => r._id) } });
     for (const reminder of remindersToRemove)
       this.container.client.reminders.delete(reminder.reminderId);
